feat: add document check to admin modification feature
