+++ UI/libs/core-admin/src/lib/components/permits/permit-detail/tabs/AttachedDocumentsTab.vue
@@ -90,6 +90,7 @@
 <script setup lang="ts">
 import SaveButton from './SaveButton.vue'
 import { UploadedDocType } from '@shared-utils/types/defaultTypes'
+import { openPdf } from '@core-admin/components/composables/openDocuments'
 import { useDocumentsStore } from '@core-admin/stores/documentsStore'
 import { usePermitsStore } from '@core-admin/stores/permitsStore'
 import { computed, reactive } from 'vue'
@@ -152,42 +153,6 @@ function onNameEdit(item, name) {
   documentStore.editApplicationFileName(oldNameWithId, newName)
 
   permitStore.updatePermitDetailApi('updated name')
-}
-
-async function openPdf(item) {
-  documentStore
-    .getUserDocument(item.name)
-    .then(response => {
-      if (response.type === 'application/pdf') {
-        const pdfBlob = new Blob([response], { type: 'application/pdf' })
-        // eslint-disable-next-line node/no-unsupported-features/node-builtins
-        const pdfUrl = URL.createObjectURL(pdfBlob)
-        const newWindow = window.open(pdfUrl, '_blank')
-
-        if (newWindow) {
-          // eslint-disable-next-line node/no-unsupported-features/node-builtins
-          URL.revokeObjectURL(pdfUrl)
-        } else {
-          alert(
-            'The PDF could not be opened in a new window. Please check your pop-up blocker settings.'
-          )
-        }
-      } else if (response.type === 'text/plain') {
-        response.text().then(base64String => {
-          fetch(base64String)
-            .then(res => res.blob())
-            .then(blob => {
-              // eslint-disable-next-line node/no-unsupported-features/node-builtins
-              const imgUrl = URL.createObjectURL(blob)
-
-              window.open(imgUrl, '_blank')
-            })
-        })
-      }
-    })
-    .catch(error => {
-      console.error('Error fetching the PDF:', error)
-    })
 }
 
 async function deletePdf() {
